MetricsManager/MetricsManagerTests/AgentsControllerUnitTests.cs
using MetricsManager;
using MetricsManager.Controllers;
using MetricsManager.Models;
using MetricsManager.Services.Interfaces;
using Microsoft.AspNetCore.Mvc;
using Microsoft.Extensions.Logging;
using Moq;
using System;
using Xunit;

namespace MetricsManagerTests
{
    public class AgentsControllerUnitTests
    {
        private AgentsController controller;
        private Mock<ILogger<AgentsController>> mockLogger;
        private Mock<IAgentsRepository> mock;
        public AgentsControllerUnitTests()
        {
            mockLogger = new Mock<ILogger<AgentsController>>();
            controller = new AgentsController(mockLogger.Object, mock.Object);
        }

        [Fact]
        public void RegisterAgent_ReturnsOk()
        {
            var agentUrl = "fasfas";

            var result = controller.RegisterAgent(agentUrl);

            _ = Assert.IsAssignableFrom<IActionResult>(result);
        }

        [Fact]
        public void EnableAgentById_ReturnsOk()
        {
            var agentId = 1;

            var result = controller.EnableAgentById(agentId);

            _ = Assert.IsAssignableFrom<IActionResult>(result);
        }

        [Fact]
        public void DisableAgentById_ReturnsOk()
        {
            var agentId = 1;

            var result = controller.DisableAgentById(agentId);

            _ = Assert.IsAssignableFrom<IActionResult>(result);
        }

        [Fact]
        public void GetRegisterMetrics_ReturnsOk()
        {
            var result = controller.EnableAgentById(4);

            Assert.Null(result);
        }
    }
}
